+++ bodhi/server/notifications.py
@@ -12,16 +12,21 @@
 # along with this program; if not, write to the Free Software
 # Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
 
-import copy
+import collections
+import logging
 import socket
 
+from sqlalchemy import event
 import fedmsg
 import fedmsg.config
 import fedmsg.encoding
-import transaction
 
+from bodhi.server import Session
 import bodhi.server
 import bodhi.server.config
+
+
+_log = logging.getLogger(__name__)
 
 
 def init(active=None, cert_prefix=None):
@@ -44,6 +49,34 @@ def init(active=None, cert_prefix=None):
 
     fedmsg.init(**fedmsg_config)
     bodhi.server.log.info("fedmsg initialized")
+
+
+@event.listens_for(Session, 'after_commit')
+def send_fedmsgs_after_commit(session):
+    """
+    An SQLAlchemy event listener to send fedmsgs after a database commit.
+
+    This relies on the session ``info`` dictionary being populated. At the moment,
+    this is done by calling the :func:`publish` function. In the future it should
+    be done automatically using SQLAlchemy listeners.
+
+    Args:
+        session (sqlalchemy.orm.session.Session): The session that was committed.
+    """
+    if 'fedmsg' in session.info:
+        # Initialize right before we try to publish, but only if we haven't
+        # initialized for this thread already.
+        if not fedmsg_is_initialized():
+            init()
+
+        for topic, messages in session.info['fedmsg'].items():
+            _log.info('emitting {n} fedmsgs to the "{topic}" topic.'.format(
+                n=len(messages), topic=topic))
+            for msg in messages:
+                fedmsg.publish(topic=topic, msg=msg)
+            # Tidy up after ourselves so a second call to commit on this session won't
+            # send the same messages again.
+            del session.info['fedmsg'][topic]
 
 
 def publish(topic, msg, force=False):
@@ -70,9 +103,13 @@ def publish(topic, msg, force=False):
         bodhi.server.log.debug("fedmsg skipping transaction and sending %r" % topic)
         fedmsg.publish(topic=topic, msg=msg)
     else:
+        # This gives us the thread-local session which we'll use to stash the fedmsg.
+        # When commit is called on it, the :func:`send_fedmsgs_after_commit` is triggered.
+        session = Session()
+        if 'fedmsg' not in session.info:
+            session.info['fedmsg'] = collections.defaultdict(list)
+        session.info['fedmsg'][topic].append(msg)
         bodhi.server.log.debug("fedmsg enqueueing %r" % topic)
-        manager = _managers_map.get_current_data_manager()
-        manager.enqueue(topic, msg)
 
 
 def fedmsg_is_initialized():
@@ -83,122 +120,3 @@ def fedmsg_is_initialized():
     # Ensure that fedmsg has an endpoint to publish to.
     context = getattr(local, '__context')
     return hasattr(context, 'publisher')
-
-
-class ManagerMapping(object):
-    """ Maintain a two-way one-to-one mapping between transaction managers and
-    data managers (for different wsgi threads in the same process). """
-
-    def __init__(self):
-        self._left = {}
-        self._right = {}
-
-    def get_current_data_manager(self):
-        current_transaction_manager = transaction.get()
-        if current_transaction_manager not in self:
-            current_data_manager = FedmsgDataManager()
-            self.add(current_transaction_manager, current_data_manager)
-            current_transaction_manager.join(current_data_manager)
-        else:
-            current_data_manager = self.get(current_transaction_manager)
-        return current_data_manager
-
-    def add(self, transaction_manager, data_manager):
-        self._left[transaction_manager] = data_manager
-        self._right[data_manager] = transaction_manager
-
-    def __contains__(self, item):
-        return item in self._left or item in self._right
-
-    def get(self, transaction_manager):
-        return self._left[transaction_manager]
-
-    def remove(self, data_manager):
-        transaction_manager = self._right[data_manager]
-        del self._left[transaction_manager]
-        del self._right[data_manager]
-
-    def __repr__(self):
-        return "<ManagerMapping: left(%i) right(%i)>" % (
-            len(self._left),
-            len(self._right),
-        )
-
-
-# This is a global object we'll maintain to keep track of the relationship
-# between transaction managers and our data managers.  It ensures that we don't
-# create multiple data managers per transaction and that we don't join the same
-# data manager to a transaction multiple times.  Our data manager should clean
-# up after itself and remove old tm/dm pairs from this mapping in the event of
-# abort or commit.
-_managers_map = ManagerMapping()
-
-
-class FedmsgDataManager(object):
-    transaction_manager = transaction.manager
-
-    def __init__(self):
-        self.uncommitted = []
-        self.committed = []
-
-    def enqueue(self, topic, msg):
-        self.uncommitted.append((topic, msg,))
-
-    def __repr__(self):
-        return self.uncommitted.__repr__()
-
-    def abort(self, transaction):
-        self.uncommitted = copy.copy(self.committed)
-        if self in _managers_map:
-            _managers_map.remove(self)
-
-    def tpc_begin(self, transaction):
-        pass
-
-    def commit(self, transaction):
-        pass
-
-    def tpc_vote(self, transaction):
-        # This ensures two things:
-        # 1) that all the objects we're about to publish are JSONifiable.
-        # 2) that we convert them from sqlalchemy objects to dicts *before* the
-        #    transaction enters its final phase, at which point our objects
-        #    will be detached from their session.
-        self.uncommitted = [
-            (topic, fedmsg.encoding.loads(fedmsg.encoding.dumps(msg)))
-            for topic, msg in self.uncommitted
-        ]
-
-        # Ensure that fedmsg has already been initialized.
-        assert fedmsg_is_initialized(), "fedmsg is not initialized"
-
-    def tpc_abort(self, transaction):
-        self.abort(transaction)
-        self._finish('aborted')
-
-    def tpc_finish(self, transaction):
-        for topic, msg in self.uncommitted:
-            bodhi.server.log.debug("fedmsg sending %r" % topic)
-            fedmsg.publish(topic=topic, msg=msg)
-        self.committed = copy.copy(self.uncommitted)
-        _managers_map.remove(self)
-        self._finish('committed')
-
-    def _finish(self, state):
-        self.state = state
-
-    def sortKey(self):
-        """ Use a 'z' to make fedmsg come last, after the db is done. """
-        return 'z_fedmsgdm' + str(id(self))
-
-    def savepoint(self):
-        return FedmsgSavepoint(self)
-
-
-class FedmsgSavepoint(object):
-    def __init__(self, dm):
-        self.dm = dm
-        self.saved_committed = copy.copy(self.dm.uncommitted)
-
-    def rollback(self):
-        self.dm.uncommitted = copy.copy(self.saved_committed)